music changed, structure of files changed
