+++ music.py
@@ -10,10 +10,15 @@ class Music:
         self.sounds = {
             'hit': pygame.mixer.Sound('Sounds/tuk.mp3'),
             'button_down': pygame.mixer.Sound('Sounds/tuk.mp3'),
+            'next_window':  pygame.mixer.Sound('Sounds/WindowChange.mp3'),
+            'defeat': pygame.mixer.Sound('Sounds/Defeat.mp3'),
+            'fail': pygame.mixer.Sound('Sounds/Fail.mp3')
         }
         self.musics = {
-            'main': 'Sounds/menu_music.mp3',
-            'game': 'Sounds/melody.mp3',
+            'main': 'Sounds/menu_music1.mp3',
+            'game': 'Sounds/game_music1.mp3',
+            'defeat': 'Sounds/Defeat.mp3',
+
         }
         self.now_play = ''
